fix(polysplit): Fix issue in print_recurse and return holes better
This change should not break anything upstream. It includes a small tweak to ensure the print_recurse variable is passed to the function that actually does the recursion.

It also ensures that polygons with holes get returned as a list of Polygon2D instead of destructively merging the hole into the boundary.