feat(polygraph): Add directed graph that can handle bad topology
This version of the directed graph can still produce cycle polygons even when the topology of the straight skeleton is not correct.

It just needs some extra methods to guide it in the direction of counterclockwise searching and then it should be able to reliably give core/perimeter polygons that can be resolved using boolean methods if they overlap.